src/backend/webhook_api/requirements.txt
Flask==2.2.3
flask-cors==3.0.10
flask-jwt-extended==4.4.4
marshmallow==3.19.0
python-dotenv==1.0.0
jsonschema==4.17.3
pytest==7.3.1
requests==2.28.2
loguru==0.7.0 